feat: Generate custom theme when user styles are saved
This commit adds listeners for saving and closing the user styles
document. When closed, it destroys the listeners. When styles are
saved, it generates the custom theme. I may need to install as a
real extension to test, because right now Nova doesn't pick up the
theme saved in 'Application Support'.