feat: Generate swatch file for all (hex) colors in theme
Now after selecting a source theme, the theme file is read and all
colors are added to a generated swatch file. This should make it a
little easier for users to reuse the theme colors when making
modifications. Right now, only hex colors are parsed, because the
swatch file expects hex colors and I'm not planning to offer
translations from other formats.